LayoutTests/media/track/track-cues-cuechange.html

<!DOCTYPE html>
<html>
    <head>
        <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />

        <script src=../media-file.js></script>
        <script src=../video-test.js></script>
    </head>
    <body>
        <p>Tests that TextTrack's cues are indexed and updated within 500ms during video playback. Test uses the cuechange event.</p>
        <p>This test may fail occasionally since we cannot guarantee that cue events will fire within 250ms of their startTime or endTime.</p>
        <video controls>
            <track id="testTrack" src="captions-webvtt/cues-chrono-order.vtt" kind="captions">
        </video>
        <script>            

        var timeInterval = 0.5;
        var cueChangeCount = 0;
        var testTrack;

        function cueChangedFromTrackElement()
        {
            consoleWrite("EVENT(cuechange) from &lt;track&gt;");
            checkCue();

            ++cueChangeCount;
            if (cueChangeCount == testTrack.track.cues.length * 2)
                endTest();
        }

        function cueChangedFromTextTrack()
        {
            consoleWrite("EVENT(cuechange) from TextTrack");
            checkCue();
        }
        
        function checkCue()
        {
            var cues = testTrack.track.cues;
            
            var currentCue = cues[Math.floor(cueChangeCount/2)];

            if (cueChangeCount % 2 == 0) {
                testExpected(currentCue, testTrack.track.activeCues[0]);
                consoleWrite("Start time:");
                testExpected("video.currentTime", currentCue.startTime, '>=');
                testExpected("video.currentTime", currentCue.startTime + timeInterval, '<');
            } else {
                consoleWrite("End time:");
                testExpected("video.currentTime", currentCue.endTime, '>=');
                testExpected("video.currentTime", currentCue.endTime + timeInterval, '<');
            }
            consoleWrite("");
        }

        waitForEvent('canplaythrough',
            function () 
            {
                testTrack = document.getElementById("testTrack");
                testExpected("testTrack.track.cues.length", 3);
                testTrack.addEventListener('cuechange', cueChangedFromTrackElement);
                testTrack.track.addEventListener('cuechange', cueChangedFromTextTrack);
                run("video.play()");
                consoleWrite("");
            }
        );

        findMediaElement();
        video.src = findMediaFile("video", "../content/test");
        
        </script>
    </body>
</html>
