LayoutTests/media/track/track-cues-cuechange-expected.txt

Tests that TextTrack's cues are indexed and updated within 500ms during video playback. Test uses the cuechange event.

This test may fail occasionally since we cannot guarantee that cue events will fire within 250ms of their startTime or endTime.

EVENT(canplaythrough)
EXPECTED (testTrack.track.cues.length == '3') OK
RUN(video.play())

EVENT(cuechange) from TextTrack
EXPECTED ([object TextTrackCue] == '[object TextTrackCue]') OK
Start time:
EXPECTED (video.currentTime >= '1') OK
EXPECTED (video.currentTime < '1.5') OK

EVENT(cuechange) from <track>
EXPECTED ([object TextTrackCue] == '[object TextTrackCue]') OK
Start time:
EXPECTED (video.currentTime >= '1') OK
EXPECTED (video.currentTime < '1.5') OK

EVENT(cuechange) from TextTrack
End time:
EXPECTED (video.currentTime >= '2') OK
EXPECTED (video.currentTime < '2.5') OK

EVENT(cuechange) from <track>
End time:
EXPECTED (video.currentTime >= '2') OK
EXPECTED (video.currentTime < '2.5') OK

EVENT(cuechange) from TextTrack
EXPECTED ([object TextTrackCue] == '[object TextTrackCue]') OK
Start time:
EXPECTED (video.currentTime >= '2.5') OK
EXPECTED (video.currentTime < '3') OK

EVENT(cuechange) from <track>
EXPECTED ([object TextTrackCue] == '[object TextTrackCue]') OK
Start time:
EXPECTED (video.currentTime >= '2.5') OK
EXPECTED (video.currentTime < '3') OK

EVENT(cuechange) from TextTrack
End time:
EXPECTED (video.currentTime >= '3.5') OK
EXPECTED (video.currentTime < '4') OK

EVENT(cuechange) from <track>
End time:
EXPECTED (video.currentTime >= '3.5') OK
EXPECTED (video.currentTime < '4') OK

EVENT(cuechange) from TextTrack
EXPECTED ([object TextTrackCue] == '[object TextTrackCue]') OK
Start time:
EXPECTED (video.currentTime >= '4') OK
EXPECTED (video.currentTime < '4.5') OK

EVENT(cuechange) from <track>
EXPECTED ([object TextTrackCue] == '[object TextTrackCue]') OK
Start time:
EXPECTED (video.currentTime >= '4') OK
EXPECTED (video.currentTime < '4.5') OK

EVENT(cuechange) from TextTrack
End time:
EXPECTED (video.currentTime >= '5') OK
EXPECTED (video.currentTime < '5.5') OK

EVENT(cuechange) from <track>
End time:
EXPECTED (video.currentTime >= '5') OK
EXPECTED (video.currentTime < '5.5') OK

END OF TEST
